Finished overhaul one.
New features:
Added "Offset calibration" -- Starts at the starting point of the motor. (be warned, might be moved or overstepped!).
Added "Calibration service" -- Calibrate the robot even when it has been running for a while.
Added "Calibration key" -- Press C to calibrate in the keyboardControlNode.
Fixed "moveToWithin" -- Calculates all speeds for the motors to have the same duration.

To do:
DeltaRobotRotation could be replaced with a fixed 3 size array.

Tests:
Movement of ((0,0,-244), 100) -> ((-20,0,-210), 100) is a nice movement. Robot moves smooth.