Changed some code to print more clear debugging info.
Changed JadeAgentX so that it will spawn new PA's when it receives a message. This enables us to activate a sniffer before the system starts.
Commented out some code in the PA's NegotiatorBehaviour to ignore EqA 2 and 3. Output of just one EqA will be more clear.
I also identified a possible cause of the messages-not-arriving-bug. It is possible that some messages are sent in rapid succession causing them not to trigger any behaviour in the receiving agent. I've started an experiment with multiple behaviours for sending and receiving multiple messages. Another solution solution would be to handle all messages in the queue using a loop in a single behaviour.