Fixed a bug in ReceiveBehaviour that caused a nullpointer exception when the behaviour is removed in the handel() method.
Changed the EA to add its own AID to the argument list when creating the SA. The SA needs it.
Fixed a bug in GetProductDuration of the SA to add the productStepId to all serviceStepMessage objects returned by a service object.
GetServiceDuration now properly saves the productStepDuration on the productStepBB and notifies the EA of it.