Made the class StartStep for the ontology StartStep.
Changed the creating of the behaviours of the equipletAgent.
Renamed some functions from the equipletAgent.
Replaced the timer from the equipletAgent to a new class.
Commented some stuff in the HardwareAgent and its behaviours because they did not work in the test.
Made a new module for the hardwareAgent: DeltaRobotModule.
Small bugfix in the HardwareAgent.
Edited the listening to changes on the blackboards to the serviceAgent.