Added the logisticsAgent class and a behaviour. The SA can ask the LA for the transport duration of parts for every prod. step. The JadeAgentX creates a single LA and passes its AID as an argument to all EA's that in turn give it to their SA.
Also changed the EvaluateBehaviour of the HA to read the ServiceStepMessage from the ServiceStepBlackBoard with the ObjectId in the received message and extract the serviceId. Then it gets the leading module for that service for step translation with the serviceId and asks it for the resulting equipletSteps. It adds up all the durations of those equipletSteps and stores the answer as the serviceStepDuration on the blackboard and notifies the SA of it.