Added _id fields of type ObjecId to ProductStepMessage and ServiceStepMessage to be able to identify different xxxStepMessages. These id's are mongo id's. If you leave them at null then BasicDBObjects created from them won't contain an _id field.
Added a bunch of behaviours to the SA and LA to handle the PlanStepWithLogistics request from an EA. These behaviours handle negotiating with the LA and filling in the parameter fields in de serviceStepBlackboard. For this to work the SA now maintains a mapping between service objects used to handle productSteps and conversationsId's used in the communication about those specific productSteps. The mapping between a service and convId is added during handling of the CanDoProductStep question because the service object for the discussed productStep is created there.