Parameter's value now is of type Object.
Fixed some bugs in ParameterGroup.
The schedule behaviours of the SA now pass along a ProductStepMessage instead of an array of partTypes.
Added a nextStep field to ServiceStepMessage that contains the ObjectId of the next service step that should be executed making a bunch of these a linked list.
Added a static method to ServiceStepMessage that sorts a given array of ServiceStepMessages based on their nextStep field. The first service step is the one that isn't referred to by any other service step.
Implemented updateParameters() in PickAndPlaceService.