Fixed some minor bugs in DeltaRobotModule and the HA.
Replaced Parameter and ParameterGroup by BasicDBObject because these are functionally the same.
Fixed a bug in GetServiceDuration and ServiceStepMessage where the nextStep field the newly generated serviceSteps did not get initialized breaking the linked list principle of the steps.
The SA now informs the HA and EA when it dies.