added the Position(double x, double y) and Position(double x, double y, Part relativeToPart) constructors to Position.
added the translate(int x, int y), translate(int x, int y, int z) and translate(Position position) methods to Position to easily change the coordinates.
Changed the implementation of fromBasicDBObject in Position so it makes sure that only known fields are in the BasicDBObject parameter.
The FillStepPlaceholders behaviour of the HA now only stores the updated instructionData returend by Module.fillPlaceHolders()
Renamed Module.isLeadingForSteps() to Module.isLeadingForServices() to better reflect what it does: return all services it can lead the translation for.
Changed GripperModule to return only the services it can handle (only "1" instead of "1" and "2")
Changed HA's configuration so that there is a pen attached to the delta robot instead of a gripper.
Fixed some minor bugs in the HA regarding BB subscriptions.
The HA now handles the ABORTED status of serviceSteps and builds a log of equipletSteps to store in the serviceStep. Then it deletes the equipletSteps.
PenModule now creates steps for at least 1 dot, even if the line is of lenght 0.
GetPartsInfoResponse now only stores the parameters, scheduleData and status fields instead of all the fields.
Added the service DrawLineService to be able to draw lines.
The service agent now reacts on the DELETED status of serviceStep to build a log, store it in the productStep and delete the serviceStep.